Fix submitted.  https://github.com/Exiv2/exiv2/issues/10
This is not a good fix because temporaryFile() is a bad idea and requires a lock to be thread safe.  However, the lock is in the sample code and not in the library.  I considered in v0.26 to rewrite this to use a temporary (MemIo) file and decided against that to save time.  My sin.  The correct fix usually requires less effort in the long run.