CMake: macro to generate documentation
We use now the Doxyfile.in template file with some variables that are
replaced at configuration time by absolute paths. Therefore, we can run
later the doxygen command with absolute paths from the
PROJECT_BINARY_DIR.

The CMake 'doc' target is added only if doxygen is available in the
system.