[slices] Simplify slices by removing inheritance in favor of SFINAE
The previous Slice implementation relied on inheritance to expose the correct
const and non-const methods. Unfortunately, this was broken by the transition to
C++11 which showed that this approach was a bad idea from the start.

The new implementation instead uses SFINAE to hide non-const member functions
when the slice holds const data. Since the problematic member functions are not
template functions themselves (they depend on the same template parameter as the
class itself), we cannot use classical SFINAE. Instead we use the Dummy template
parameter trick described e.g. here:
https://foonathan.net/blog/2016/12/21/conditionally-removing-functions.html