Conan: Updates to fix sudden linkage issues
We suddenly started to have some linking issues in the 0.27 branch after
the libcurl packages were updated in the conan-center repositories.
After some experimentation I took the following steps to fix up the
situation:

- Update conan to latest version
- Update of libcurl to the latest version available
- Use libcurl static libraries

Note that the change to use static libraries is just to make the
deployment step as easier as possible in the travis builds.