Fix misdetection of Sigma 100-400mm f/5-6.3 as Canon EF 100-400mm f/4.5-5.6
When searching for the Sigma lens, only the string "5" is searched in the
lens description, which also happens to be present for the Canon lens.
Since the Canon lens comes first in the list, it wins. Fix this issue by
prefixing the search with f/ so it always has to match the beginning of the
aperture specification.

This will still allow mismatches for example when searching for f/4 that
also matches f/4.5, but no such issues occur at the moment, so the simple
fix is sufficient for now.