Add custom update-pot target
While working on a separate localization bug, I noticed that
`po/exiv2.pot` hasn't been updated since 2017, and that with the removal
of the autotools build system there's no longer any way to update it.
Add a simple `update-pot` target to do so, and bring `po/POTFILES.in` up
to date with the current source tree.

I haven't updated `po/exiv2.pot` itself because that would be a large
diff, though I've checked that the output looks reasonable.  I suggest
that maintainers do this.