I'm close to giving up on this.  There are several issues:
1.  We must set conan-center to https://center.conan.io in `conan remote`
The v0.27.5 builds were successfully achieved with the 'bincrafters' recipes.

I've fixed this.

2.  We were not setting CMAKE_MODULE_PATH correctly.
It should be a semicolon separated PATH style variable.

I've fixed this.

When we set it correctly, the FindXXX.cmake generated by conan are used by the compiler and linker to the code compiles.  When this was not being set, we were falling back on the CMake provided FindXXX.cmake files.  If the dependencies were installed by the package manager, those usually work.

3.  We need to add `${GTest_INCLUDE_DIRS}` the include path in unitTests

This was working in v0.27.5 because we hadn't fixed the CMAKE_MODULE_PATH.  The compiler found gtest.h in a standard location.

I've fixed this.

4. We should default to CMAKE_CXX_STANDARD 98 and gtest/1.8.1

I've fix this.

4. There is something wrong when linking GTest.

It fails to link all platforms.   This issue was masked by issues 1,2 and 3.

For the moment, the only way to submit this PR is to ignore unit tests.  If it makes the build stable, that will do and we can ask Luis to look at the GTest linking issue.