Update ubuntu-latest tag to ubuntu-22.04
Everytime a CI service such a GitHub actions update the images
corresponding to their 'latest' tags, pipelines tend to have some
hipcups due to changes in the distribution default packages.

By fixing the images to a specific one, we should be more resilient to
changes in the Github actions environment.