XMPMeta.cpp: match types to format strings in OutProc* macros
The numerical OutProc*(num) macros fill in their arguments via
snprintf's "..." varargs part, and the type at the use site depends on
the passed-in types.  This might cause wrong types on the stack that
cause undefined behavior in the snprintf() function, and reading past
memory, outputting garbage.

static_cast<> the macro arguments to the types matching the format
string, to get the expected type width on the stack.