Fix XMPUtils::ConvertToInt64 sscanf() type
This function uses sscanf() with "%lld" or "%llx" to
parse a string to an integer. However, nothing guarantees
that sizeof(XMP_Int64) == sizeof(long long); the latter
is what sscanf will fill with either of these format strings,
to avoid memory corruption.

Use a long long to match the sscanf size, and let the return statement
cast it to XMP_Int64.