tests: fix parsing w/ path containing "-pa" or "-pS"
The OutputTagExtract test case for tiff_test runs exiv2 two times,
with "-pa" and "-pS", and parses their outputs. To know which output to
parse, it checks for "-pa" and "-pS" in the string of the command being
run; considering that the command string contains the full path to the
test data, which is a subdirectory of the sources, this means that a
wrong parser will be used in case the full source path contains any of
"-pa" or "-pS" (e.g. "/build/some-path/exiv2/...").

Cheap fix for this: since "-pa"/"-pS" are options in the command string,
check for them using spaces around.

(cherry picked from commit 5f4ffdd4830838fec582c4693e7ee74d8ece541a)