Throw if open file with EXV_ENABLE_FILESYSTEM off
If ImageFactory::createIo() is called with a file path and has been
built with EXV_ENABLE_FILESYSTEM off it returns a NULL pointer.
ImageFactory::open() then calls io->open() on this pointer, causing a
segfault. The documentation of ImageFactory::createIo does not say it
can return a NULL pointer. This commit makes it throw an exception
instead of returning a NULL pointer.