rename: path separator for new path based on old path
for concatenation of new file name, concatenation operator of std::filesystem::path is not used:
On MSYS2 UCRT64 the path separator to be used in terminal is slash, but as concatenation operator a back slash will be added. Rename works but with verbose a path with different operators will be shown.